requirements-minimal.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
WTForms>=3.1.0
Werkzeug>=3.0.0
stripe>=8.0.0
python-dotenv>=1.0.0
email-validator>=2.1.0
bcrypt>=4.1.0
tabulate>=0.9.0
flask-cors>=4.0.0
